taq-hackathon/app/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
psycopg[binary]>=3.1
python-dotenv>=1.0
